+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "br.com.ddm.hogwartshistorys"
+        applicationId "br.com.ddm.hogwartshistories"
         minSdkVersion 28
         targetSdkVersion 30
         versionCode 1
