+++ app/src/androidTest/java/br/com/ddm/hogwartshistories/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package br.com.ddm.hogwartshistorys
+package br.com.ddm.hogwartshistories
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("br.com.ddm.hogwartshistorys", appContext.packageName)
+        assertEquals("br.com.ddm.hogwartshistories", appContext.packageName)
     }
 }
