+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
-    <string name="app_name">Hogwarts Historys</string>
-    <string name="initial_message">Bem Vindo à Hogwarts!</string>
+    <string name="app_name">Hogwarts Histories</string>
+    <string name="initial_message">Bem-Vindo à Hogwarts!</string>
     <string name="action_buscar">Buscar</string>
     <string name="action_atualizar">Atualizar</string>
     <string name="grifinoria">Grifinória Rules</string>
